Extract DerivePersist into a module
